+++ integration/flink/examples/stateful/src/main/java/io/openlineage/flink/FlinkIcebergApplication.java
@@ -5,6 +5,7 @@
 
 package io.openlineage.flink;
 
+import io.openlineage.util.FlinkListenerUtils;
 import org.apache.flink.core.execution.JobListener;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
@@ -34,12 +35,7 @@ public class FlinkIcebergApplication {
       .overwrite(true)
       .append();
 
-    // we use this app to test open lineage flink integration so it cannot make use of OpenLineageFlinkJobListener classes
-    JobListener openlineageJobListener = (JobListener) Class.forName("io.openlineage.flink.OpenLineageFlinkJobListener")
-      .getConstructor(StreamExecutionEnvironment.class)
-      .newInstance(env);
-
-    env.registerJobListener(openlineageJobListener);
+    env.registerJobListener(FlinkListenerUtils.instantiate(env));
     env.execute("flink-examples-iceberg");
   }
 }