+++ integration/flink/src/main/java/io/openlineage/flink/visitor/lifecycle/FlinkExecutionContext.java
@@ -8,6 +8,7 @@ package io.openlineage.flink.visitor.lifecycle;
 import io.openlineage.client.OpenLineage;
 import io.openlineage.client.OpenLineage.RunEvent;
 import io.openlineage.client.OpenLineage.RunEvent.EventType;
+import io.openlineage.client.OpenLineage.RunEventBuilder;
 import io.openlineage.flink.SinkLineage;
 import io.openlineage.flink.TransformationUtils;
 import io.openlineage.flink.api.OpenLineageContext;
@@ -21,6 +22,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.exception.ExceptionUtils;
@@ -29,32 +31,25 @@ import org.apache.flink.api.common.JobID;
 import org.apache.flink.api.dag.Transformation;
 
 @Slf4j
+@Builder
 public class FlinkExecutionContext implements ExecutionContext {
 
   @Getter private final JobID jobId;
-
-  private final UUID runId;
-
-  private final EventEmitter eventEmitter;
-  private final OpenLineageContext openLineageContext;
+  protected final UUID runId;
+  protected final EventEmitter eventEmitter;
+  protected final OpenLineageContext openLineageContext;
+  private final String jobName;
+  private final String jobNamespace;
 
   @Getter private final List<Transformation<?>> transformations;
-
-  public FlinkExecutionContext(JobID jobId, List<Transformation<?>> transformations) {
-    this.jobId = jobId;
-    this.runId = UUID.randomUUID();
-    this.transformations = transformations;
-    this.eventEmitter = new EventEmitter();
-    this.openLineageContext =
-        OpenLineageContext.builder()
-            .openLineage(new OpenLineage(EventEmitter.OPEN_LINEAGE_CLIENT_URI))
-            .build();
-  }
 
   @Override
   public void onJobSubmitted() {
     log.debug("JobClient - jobId: {}", jobId);
-    RunEvent runEvent = buildEventForEventType(EventType.START).build();
+    RunEvent runEvent =
+        buildEventForEventType(EventType.START)
+            .run(new OpenLineage.RunBuilder().runId(runId).build())
+            .build();
     log.debug("Posting event for onJobSubmitted {}: {}", jobId, runEvent);
     eventEmitter.emit(runEvent);
   }
@@ -87,22 +82,15 @@ public class FlinkExecutionContext implements ExecutionContext {
       outputDatasets.addAll(getOutputDatasets(visitorFactory, lineage.getSink()));
     }
 
-    return openLineageContext
-        .getOpenLineage()
-        .newRunEventBuilder()
-        .inputs(inputDatasets)
-        .outputs(outputDatasets)
-        .eventType(eventType);
+    return commonEventBuilder().inputs(inputDatasets).outputs(outputDatasets).eventType(eventType);
   }
 
   @Override
   public void onJobCompleted(JobExecutionResult jobExecutionResult) {
     OpenLineage openLineage = openLineageContext.getOpenLineage();
     eventEmitter.emit(
-        openLineage
-            .newRunEventBuilder()
+        commonEventBuilder()
             .run(openLineage.newRun(runId, null))
-            .eventTime(ZonedDateTime.now())
             .eventType(EventType.COMPLETE)
             .build());
   }
@@ -111,8 +99,7 @@ public class FlinkExecutionContext implements ExecutionContext {
   public void onJobFailed(Throwable failed) {
     OpenLineage openLineage = openLineageContext.getOpenLineage();
     eventEmitter.emit(
-        openLineage
-            .newRunEventBuilder()
+        commonEventBuilder()
             .run(
                 openLineage.newRun(
                     runId,
@@ -122,9 +109,22 @@ public class FlinkExecutionContext implements ExecutionContext {
                             openLineage.newErrorMessageRunFacet(
                                 failed.getMessage(), "JAVA", ExceptionUtils.getStackTrace(failed)))
                         .build()))
-            .eventTime(ZonedDateTime.now())
             .eventType(EventType.FAIL)
+            .eventTime(ZonedDateTime.now())
             .build());
+  }
+
+  /**
+   * Builds common elements of Openlineage events constructed.
+   *
+   * @return
+   */
+  private RunEventBuilder commonEventBuilder() {
+    OpenLineage openLineage = openLineageContext.getOpenLineage();
+    return openLineage
+        .newRunEventBuilder()
+        .job(openLineage.newJob(jobNamespace, jobName, null))
+        .eventTime(ZonedDateTime.now());
   }
 
   private List<OpenLineage.InputDataset> getInputDatasets(